refactor: moved code to get lnurl details into func
